Lesson2
Refactor + start-email
