--- lesson-2/shop/.gitignore
@@ -1,1 +1,0 @@
-config-with-passwords.yaml
