+++ lesson-2/shop/go.mod
@@ -1,10 +1,8 @@
-module shop
+module gb-go-architecture/lesson-1/shop_new
 
-go 1.14
+go 1.15
 
 require (
-	github.com/go-telegram-bot-api/telegram-bot-api v4.6.4+incompatible
-	github.com/gorilla/mux v1.7.4
-	github.com/technoweenie/multipartstreamer v1.0.1 // indirect
-	gopkg.in/yaml.v2 v2.3.0
+	github.com/gorilla/mux v1.8.0
+	go.uber.org/zap v1.16.0
 )
