+++ lesson-2/shop/models/item.go
@@ -1,7 +1,11 @@
 package models
 
+import "time"
+
 type Item struct {
-	ID    int32   `json:"id"`
-	Name  string  `json:"name"`
-	Price float64 `json:"price"`
+	ID        int32     `json:"id"`
+	Name      string    `json:"name"`
+	Price     int64     `json:"price"`
+	CreatedAt time.Time `json:"created_at"`
+	UpdatedAt time.Time `json:"updated_at"`
 }
