+++ lesson-2/shop/models/order.go
@@ -1,8 +1,13 @@
 package models
 
+import "time"
+
 type Order struct {
-	ID      int32   `json:"id"`
-	Phone   string  `json:"phone"`
-	Email	string	`json:"email"`
-	ItemIDs []int32 `json:"item_ids"`
+	ID            int32     `json:"id"`
+	CustomerName  string    `json:"customer_name"`
+	CustomerPhone string    `json:"customer_phone"`
+	Email         string    `json:"email"`
+	ItemIDs       []int32   `json:"item_ids"`
+	CreatedAt     time.Time `json:"crated_at"`
+	UpdatedAt     time.Time `json:'updated_at"`
 }
