+++ lesson-2/shop/repository/filter.go
@@ -4,12 +4,8 @@ type baseFilter struct {
 	Limit  int
 	Offset int
 }
-
 type ItemFilter struct {
 	baseFilter
-
-	ItemIDs []int32
-
 	PriceLeft  *int64
 	PriceRight *int64
 }
