--- lesson-2/shop_new/go.mod
@@ -1,10 +1,0 @@
-module shop
-
-go 1.15
-
-require (
-	github.com/go-telegram-bot-api/telegram-bot-api v4.6.4+incompatible
-	github.com/gorilla/mux v1.8.0
-	github.com/technoweenie/multipartstreamer v1.0.1 // indirect
-	go.uber.org/zap v1.16.0
-)
