--- lesson-2/shop_new/models/item.go
@@ -1,11 +1,0 @@
-package models
-
-import "time"
-
-type Item struct {
-	ID        int32     `json:"id"`
-	Name      string    `json:"name"`
-	Price     int64     `json:"price"`
-	CreatedAt time.Time `json:"created_at"`
-	UpdatedAt time.Time `json:"updated_at"`
-}
